+++ lib/clickhouse-ec2-stack.js
@@ -75,7 +75,6 @@ class ClickhouseEc2Stack extends Stack {
     const userData = ec2.UserData.forLinux();
     userData.addCommands(
       fs.readFileSync(path.join(__dirname, '..', 'scripts', 'install_clickhouse.sh'), 'utf8'),
-      // `export REGION=${props.env.region}`,
       'sudo mkdir -p /opt/scripts/',
       fs.readFileSync(path.join(__dirname, '..', 'scripts', 'install_clickhouse_backup.sh'), 'utf8'),
       fs.readFileSync(path.join(__dirname, '..', 'scripts', 'clickhouse_backup.sh'), 'utf8'),
