+++ lib/dynamodb-stack.js
@@ -8,14 +8,13 @@ class DynamoDbStack extends Stack {
     super(scope, id, props);
 
     const table = new dynamodb.Table(this, 'StreamTableMap', {
-      tableName: 'tables_streams',
-      // change this ^ to stream_table_map in production
+      tableName: 'stream_table_map',
       partitionKey: { name: 'stream_id', type: dynamodb.AttributeType.STRING },
       sortKey: { name: 'table_id', type: dynamodb.AttributeType.STRING },
       billingMode: dynamodb.BillingMode.PROVISIONED,
       readCapacity: 3,
       writeCapacity: 3,
-      removalPolicy: RemovalPolicy.DESTROY, // CAUTION: This will delete the table when the stack is destroyed
+      removalPolicy: RemovalPolicy.DESTROY,
     });
 
     table.addGlobalSecondaryIndex({
