+++ lib/flask-ec2-stack.js
@@ -37,18 +37,10 @@ class FlaskEc2Stack extends Stack {
       "Allow SSH access",
     );
 
-    // const clickhouseSecurityGroup = ec2.SecurityGroup.fromSecurityGroupId(
-    //   this,
-    //   'ClickHouseSecurityGroup',
-    //   Fn.importValue('ClickHouseSecurityGroupId')
-    // );
-    // flaskSecurityGroup.addEgressRule(clickhouseSecurityGroup, ec2.Port.tcp(8123), 'Allow traffic to ClickHouse');
-
     const role = new iam.Role(this, "FlaskEC2Role", {
       assumedBy: new iam.ServicePrincipal("ec2.amazonaws.com"),
     });
 
-    // Add managed policies
     role.addManagedPolicy(
       iam.ManagedPolicy.fromAwsManagedPolicyName(
         "AmazonSSMManagedInstanceCore",
@@ -58,13 +50,10 @@ class FlaskEc2Stack extends Stack {
       iam.ManagedPolicy.fromAwsManagedPolicyName("CloudWatchAgentServerPolicy"),
     );
 
-    // Add policy for DynamoDB access
     role.addToPolicy(
       new iam.PolicyStatement({
         effect: iam.Effect.ALLOW,
         actions: ["dynamodb:*", "lambda:*"],
-        //resources: ['arn:aws:dynamodb:*:*:table/tables_streams'],
-        // note change ^ to stream_table_map for production
         resources: "*",
       }),
     );
@@ -111,12 +100,6 @@ class FlaskEc2Stack extends Stack {
       description: "Flask EC2 Instance ID",
       exportName: "FlaskInstanceId",
     });
-
-    // new CfnOutput(this, 'SecurityGroupId', {
-    //   value: flaskSecurityGroup.securityGroupId,
-    //   description: 'Security Group ID for Flask instance',
-    //   exportName: 'FlaskSecurityGroupId',
-    // });
   }
 }
