+++ lib/lambda-stack.js
@@ -32,7 +32,7 @@ class LambdaStack extends Stack {
       functionName: 'kinesis-to-clickhouse-dev',
       runtime: lambda.Runtime.PYTHON_3_12,
       handler: 'lambda_function.lambda_handler',
-      code: lambda.Code.fromAsset('lambda_function.zip'),
+      code: lambda.Code.fromAsset('lambda/lambda_function.zip'),
       role: lambdaRole,
       timeout: Duration.seconds(300),
       environment: {
@@ -42,7 +42,7 @@ class LambdaStack extends Stack {
     });
 
     const lambdaLayer = new lambda.LayerVersion(this, 'LambdaLayer', {
-      code: lambda.Code.fromAsset('layer_content.zip'),
+      code: lambda.Code.fromAsset('layers/layer_content.zip'),
       compatibleRuntimes: [lambda.Runtime.PYTHON_3_12],
       description: 'Layer for Kinesis to ClickHouse Lambda function',
     });
