+++ src/cli/set-profile.js
@@ -2,7 +2,6 @@ const { input } = require('@inquirer/prompts');
 
 
 async function setProfile() {
-  return 'capstone-team4'
   const answer = await input({message: 'Enter your AWS Profile name:'});
   return answer;
 }
